fix(Fix-B-011): fix training_histories migration error duration old training Pelatihan Teknis #2

+++ database/migrations/2024_01_01_000024_create_training_histories_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->year('period_year')->nullable();
             $table->date('start_date')->nullable();
             $table->date('end_date')->nullable();
-            $table->smallInteger('duration')->nullable()->comment('in hours');
+            $table->integer('duration')->nullable()->comment('in hours');
             $table->string('organizer', 512)->nullable();
             $table->string('reference_number', 160)->nullable();
             $table->text('link')->nullable();
